requirements.txt
fastmcp>=0.1.0
flask>=2.0.0
flask-cors>=4.0.0
flask-limiter>=3.0.0
flask-caching>=2.0.0
flask-compress>=1.0.0
sqlalchemy>=2.0.0
gunicorn>=21.0.0
redis>=4.0.0
flask-restx>=1.0.0
apispec>=6.0.0
marshmallow>=3.0.0
pyjwt>=2.0.0
bcrypt>=4.0.0
flask-jwt-extended>=4.0.0
